refactor: Separate decoder_layer and decoder
